Added OOPS

+++ Lab 4/src/exercise1/Person.java
@@ -3,11 +3,11 @@ package exercise1;
 public class Person {
 	
 	private String name;
-	private float age;
+	private int age;
 	
-	public Person(String name, float age) {
+	public Person(String name, int age) {
 		this.name= name;
-		this.age= age;
+		this.age=age;
 	}
 	
 	//toString method
@@ -27,7 +27,7 @@ public class Person {
 	public float getAge() {
 		return age;
 	}
-	public void setAge(float age) {
+	public void setAge(int age) {
 		this.age = age;
 	}
